+++ src/deluge/gui/menu_item/midi/devices.h
@@ -19,8 +19,6 @@
 
 #include "gui/menu_item/value.h"
 
-class MIDIDevice;
-
 namespace deluge::gui::menu_item::midi {
 
 class Devices final : public Value<int32_t> {
@@ -28,7 +26,7 @@ public:
 	using Value::Value;
 	void beginSession(MenuItem* navigatedBackwardFrom = nullptr) override;
 	void selectEncoderAction(int32_t offset) override;
-	MIDIDevice* getDevice(int32_t deviceIndex);
+	MIDICable* getCable(int32_t deviceIndex);
 	virtual void drawValue();
 	MenuItem* selectButtonPress() override;
 	void drawPixelsForOled();