+++ src/deluge/model/drum/drum.cpp
@@ -86,15 +86,15 @@ void Drum::getCombinedExpressionInputs(int16_t* combined) {
 }
 
 void Drum::expressionEventPossiblyToRecord(ModelStackWithTimelineCounter* modelStack, int16_t newValue,
-                                           int32_t whichExpressionimension, int32_t level) {
+                                           int32_t expressionDimension, int32_t level) {
 
 	// Ok we have to first combine the expression inputs that the user might have sent at both MPE/polyphonic/finger
 	// level, *and* at channel/instrument level. Yes, we combine these here at the input before the data gets recorded
 	// or sounded, because unlike for Instruments, we're a Drum, and all we have is the NoteRow level to store this
 	// stuff.
-	lastExpressionInputsReceived[level][whichExpressionimension] = newValue >> 8; // Store value
+	lastExpressionInputsReceived[level][expressionDimension] = newValue >> 8; // Store value
 	int32_t combinedValue =
-	    (int32_t)newValue + ((int32_t)lastExpressionInputsReceived[!level][whichExpressionimension] << 8);
+	    (int32_t)newValue + ((int32_t)lastExpressionInputsReceived[!level][expressionDimension] << 8);
 	combinedValue = lshiftAndSaturate<16>(combinedValue);
 
 	expressionValueChangesMustBeDoneSmoothly = true;
@@ -111,8 +111,8 @@ void Drum::expressionEventPossiblyToRecord(ModelStackWithTimelineCounter* modelS
 			goto justSend;
 		}
 
-		bool success = noteRow->recordPolyphonicExpressionEvent(modelStackWithNoteRow, combinedValue,
-		                                                        whichExpressionimension, true);
+		bool success =
+		    noteRow->recordPolyphonicExpressionEvent(modelStackWithNoteRow, combinedValue, expressionDimension, true);
 		if (!success) {
 			goto justSend;
 		}
@@ -121,7 +121,7 @@ void Drum::expressionEventPossiblyToRecord(ModelStackWithTimelineCounter* modelS
 	// Or if not recording, just sound the change ourselves here (as opposed to the AutoParam doing it).
 	else {
 justSend:
-		expressionEvent(combinedValue, whichExpressionimension); // Virtual function
+		expressionEvent(combinedValue, expressionDimension); // Virtual function
 	}
 
 	expressionValueChangesMustBeDoneSmoothly = false;